move readFile & writeFile Func, OutputFile Struct in File Struct
