+++ JSONParser/JSONParser/InputView.swift
@@ -24,17 +24,4 @@ struct InputView {
         }
         return inputValue
     }
-    
-    public static func readFile(inputFile: String) throws -> String {
-        do {
-            let fileManager = FileManager()
-            var path = try fileManager.url(for: .documentDirectory, in: .userDomainMask, appropriateFor: nil, create: false)
-            path.appendPathComponent(inputFile)
-            let contents = try String(contentsOf: path)
-            return contents
-        } catch {
-            throw JsonError.fileNotFound
-        }
-    }
-    
 }
