+++ JSONParser/JSONParser/JsonError.swift
@@ -13,6 +13,7 @@ enum JsonError : Error {
     case unSupportedObjectPattern
     case fileNotFound
     case unableToCreateFile
+    case pathNotFound
     case unknown
     
     func description() -> String {
@@ -25,6 +26,8 @@ enum JsonError : Error {
             return "해당 파일을 찾을 수가 없습니다."
         case .unableToCreateFile:
             return "파일을 생성할 수 없습니다."
+        case .pathNotFound:
+            return "경로를 찾을 수 없습니다."
         case .unknown:
             return "알 수 없는 에러가 발생하였습니다."
         }
